Move ray_T_ij outside of kernels, as a separate API module.
